refactor(components): [checkbox]split hooks into different files
